Add CorsFilterTest - (Task #952)
Extracted from ModelAccessResourceTest.

---
Task #952: Refactor ModelAccessResource
